resume_ranker/requirements.txt
Django>=5.2,<6.0
django-environ>=0.11
openai>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
tenacity>=8.2
celery>=5.3
redis>=5.0
gunicorn>=21.2
whitenoise>=6.6
